feat : 9 Menu routes (#9)

+++ controllers/order.controller.js
@@ -1,6 +1,7 @@
 const { orderService } = require("../service/order.service");
 const { orderSchema } = require("../validator/order.schema.validator");
 const mapRequestBodyToOrder = require("../utils/mapper");
+const mongoose = require("mongoose");
 
 const orderController = {
   async createOrder(req, res) {
@@ -14,6 +15,7 @@ const orderController = {
       res.status(400).json({ error: error.message });
     }
   },
+
   async getOrders(req, res) {
     try {
       const orders = await orderService.getOrders();
@@ -22,33 +24,43 @@ const orderController = {
       res.status(500).json({ error: "Internal server error" });
     }
   },
+
   async updateOrderStatus(req, res) {
     try {
       const { orderId } = req.params;
       const { Status } = req.body;
-      if (Status == "CANCEL_CONFIRMED") {
+
+      if (!mongoose.Types.ObjectId.isValid(orderId)) {
+        return res.status(400).json({ error: "Invalid Order ID" });
+      }
+
+      if (Status === "CANCEL_CONFIRMED") {
         await orderService.deleteOrder(orderId);
-        res.json({ msg: "order canceled" });
+        res.json({ msg: "Order canceled" });
       } else {
-        const updatedOrder = await orderService.updateOrderStatus(
-          orderId,
-          Status
-        );
+        const updatedOrder = await orderService.updateOrderStatus(orderId, Status);
         res.json(updatedOrder);
       }
     } catch (error) {
       res.status(400).json({ error: error.message });
     }
   },
+
   async deleteOrder(req, res) {
     try {
       const { orderId } = req.params;
+
+      if (!mongoose.Types.ObjectId.isValid(orderId)) {
+        return res.status(400).json({ error: "Invalid Order ID" });
+      }
+
       const deletedOrder = await orderService.deleteOrder(orderId);
       res.json(deletedOrder);
     } catch (error) {
       res.status(400).json({ error: error.message });
     }
   },
+
   async getOrderByTableId(req, res) {
     try {
       const { tableId } = req.params;
@@ -67,12 +79,12 @@ const orderController = {
       res.status(500).json({ error: "Internal server error" });
     }
   },
+
   async getOrdersByMonth(req, res) {
     const { month, year } = req.params; // Assuming month and year are passed as route parameters
 
     try {
       const orders = await orderService.getOrdersByMonth(month, year);
-
       res.json({ orders });
     } catch (error) {
       console.error("Error fetching orders by month:", error);
